backend/requirements.txt
# Core Framework - Python 3.13 compatible versions
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1

# Document Processing
PyMuPDF==1.26.7
python-docx==1.1.2
pdfplumber==0.11.6

# Vector Database - Daha yeni versiyon Python 3.13 için
chromadb==0.5.23

# Google Gemini
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0

# Logging
python-json-logger==3.2.1